Refactor tabs as within data panel (instead of on toolbar)
