--- web/shapeworks/src/views/Analyze.vue
@@ -1,23 +1,0 @@
-<template>
-  <div>Analyze</div>
-</template>
-
-<script lang="ts">
-import { loadDataset } from "@/store";
-import { defineComponent, onMounted } from "@vue/composition-api";
-
-export default defineComponent({
-  props: {
-      dataset: {
-          type: Number,
-          required: true,
-      },
-
-  },
-  setup(props) {
-    onMounted(async () => {
-      await loadDataset(props.dataset);
-    });
-  }
-});
-</script>
