--- web/shapeworks/src/views/Data.vue
@@ -1,313 +1,0 @@
-<script lang="ts">
-import imageReader from '../reader/image';
-import { groupBy, shortDateString, shortFileName } from '../helper';
-import { getDataObjectsForSubject } from '@/api/rest';
-import { defineComponent, onMounted, ref, watch } from '@vue/composition-api';
-import { DataObject, ShapeData, Subject } from '../types';
-import { getSubjectsForDataset } from '@/api/rest'
-import ShapeViewer from '../components/ShapeViewer.vue';
-import {
-    selectedDataset,
-    loadDataset,
-    allDataObjectsInDataset,
-    selectedDataObjects,
-    loadingState,
-} from '../store';
-import router from '@/router';
-import vtkPolyData from 'vtk.js/Sources/Common/DataModel/PolyData';
-
-
-export default defineComponent({
-    components: {
-        ShapeViewer,
-    },
-    props: {
-        dataset: {
-            type: Number,
-            required: true,
-        },
-    },
-    setup(props) {
-        const mini = ref(false);
-        const anatomies = ref<string[]>([]);
-        const selectedAnatomies = ref<string[]>([]);
-        const subjects = ref<Record<string, Subject>>({})
-        const selectedSubjects = ref<number[]>([])
-        const rows = ref<number>(1);
-        const cols = ref<number>(1);
-        const renderData = ref<Record<string, ShapeData[]>>({});
-        const headers =  [
-            {text: 'ID', value: 'id'},
-            {text: 'Anatomy', value: 'anatomy_type'},
-            {text: 'Type', value: 'type'},
-            {text: 'File Name', value: 'file', cellClass: 'file-column'},
-        ];
-
-        async function fetchData(datasetId: number) {
-            loadingState.value = true;
-            subjects.value = Object.fromEntries((await getSubjectsForDataset(datasetId)).sort((a, b) => {
-                if(a.created < b.created) return 1;
-                if(a.created > b.created) return -1;
-                return 0;
-            }).map((subject: Subject) => [subject.id, subject]));
-            allDataObjectsInDataset.value =  (await Promise.all(
-                Object.values(subjects.value).map(
-                    async (subject: Subject) => (await getDataObjectsForSubject(subject.id)).map(
-                        (dataObj) => {
-                            let assignments = {'uid': `${dataObj.type}_${dataObj.id}`}
-                            // TODO: We use anatomy_type and modality interchangeably
-                            if(!dataObj.anatomy_type && dataObj.modality){
-                                assignments = Object.assign(assignments, {'anatomy_type': dataObj.modality})
-                            }
-                            return Object.assign(dataObj, assignments)
-                        }
-                    )
-                )
-            )).flat()
-            anatomies.value = Object.keys(groupBy(allDataObjectsInDataset.value, 'anatomy_type'))
-            loadingState.value = false;
-        }
-
-        function updateSelectedObjects(){
-            selectedDataObjects.value = allDataObjectsInDataset.value.filter(
-                (dataObject) => {
-                    return selectedAnatomies.value.includes(dataObject.anatomy_type) &&
-                    selectedSubjects.value.includes(dataObject.subject)
-                }
-            )
-        }
-
-        onMounted(async () => {
-            await loadDataset(props.dataset);
-            if(!selectedDataset.value) {
-                router.push({
-                    name: 'select',
-                });
-                return;
-            }
-            await fetchData(selectedDataset.value.id)
-        })
-
-        watch(selectedAnatomies, updateSelectedObjects)
-
-        watch(selectedSubjects, updateSelectedObjects)
-
-        watch(selectedDataObjects, async () => {
-            renderData.value = {}
-            const groupedSelections: Record<string, DataObject[]> = groupBy(selectedDataObjects.value, 'subject')
-            renderData.value = Object.fromEntries(
-                await Promise.all(Object.entries(groupedSelections).map(
-                    async ([subjectId, dataObjects]) => [
-                        subjects.value[subjectId].name,
-                        (await Promise.all(dataObjects.map(
-                            (dataObject) => imageReader(
-                                dataObject.file,
-                                shortFileName(dataObject.file)
-                            )
-                        ))).map((imageData) => ({shape: imageData, points: vtkPolyData.newInstance()}))
-                    ]
-                )
-            ))
-
-            const n = Object.keys(renderData.value).length;
-            const sqrt = Math.ceil(Math.sqrt(n));
-            if(sqrt <= 5) {
-                rows.value = Math.ceil(n / sqrt);
-                cols.value = sqrt;
-            } else {
-                rows.value = Math.ceil(n / 5);
-                cols.value = 5;
-            }
-        })
-
-        return {
-            mini,
-            anatomies,
-            selectedAnatomies,
-            subjects,
-            selectedSubjects,
-            rows,
-            cols,
-            renderData,
-            headers,
-            selectedDataset,
-            allDataObjectsInDataset,
-            selectedDataObjects,
-            shortFileName,
-            shortDateString,
-        }
-    }
-})
-</script>
-
-
-<template>
-    <div style="height: 100%">
-        <div class="context-card">
-            <div>
-                <div class="text-overline">
-                    DATASET ({{ selectedDataset.created.split('T')[0] }})
-                </div>
-                <v-list-item-title class="text-h6 mb-1">
-                    {{ selectedDataset.name }}
-                </v-list-item-title>
-            </div>
-        </div>
-        <v-divider />
-
-        <div class='content-area'>
-            <v-navigation-drawer :mini-variant.sync="mini" width="500" absolute>
-                <v-list-item>
-                    <v-btn
-                        icon
-                        @click.stop="mini=false"
-                        class="mr-3"
-                    >
-                    <v-icon large>mdi-database</v-icon>
-                    </v-btn>
-                    <v-list-item-title class="text-h6">
-                        Data Objects ({{ allDataObjectsInDataset.length }})
-                    </v-list-item-title>
-                    <v-btn
-                        icon
-                        v-if="!mini"
-                        @click.stop="mini=true"
-                    >
-                    <v-icon large>mdi-chevron-left</v-icon>
-                    </v-btn>
-                </v-list-item>
-                <v-list-item>
-                    <v-icon />
-                    <div style="width:100%">
-                        <v-list dense shaped>
-                            <v-subheader>ANATOMIES</v-subheader>
-                            <v-list-item-group>
-                                <v-list-item
-                                    v-for="anatomy_type in anatomies"
-                                    :key="anatomy_type"
-                                >
-                                     <v-checkbox
-                                        color="primary"
-                                        v-model="selectedAnatomies"
-                                        :value="anatomy_type"
-                                    ></v-checkbox>
-                                    <v-list-item-title v-text="anatomy_type"/>
-                                </v-list-item>
-                            </v-list-item-group>
-                            <v-subheader>SUBJECTS</v-subheader>
-                            <v-list-item-group>
-                                <v-list-group
-                                    v-for="subject in Object.values(subjects)"
-                                    :key="subject.id"
-                                    v-model="subject.showDetails"
-                                >
-                                    <template v-slot:activator>
-                                        <v-checkbox
-                                            color="primary"
-                                            v-model="selectedSubjects"
-                                            :value="subject.id"
-                                            @click.native="$event.stopPropagation()"
-                                        ></v-checkbox>
-                                        <v-list-item-title>
-                                            Subject: {{ subject.name }}
-                                        </v-list-item-title>
-                                    </template>
-
-                                    <v-data-table
-                                        :headers="headers"
-                                        :items="allDataObjectsInDataset.filter((obj) => obj.subject === subject.id)"
-                                        item-key="uid"
-                                        disable-pagination
-                                        hide-default-footer
-                                        dense
-                                        width="100%"
-                                    >
-                                        <!-- eslint-disable-next-line -->
-                                        <template v-slot:group.header="{ group, headers, toggle, isOpen }">
-                                            <td :colspan="headers.length">
-                                                <v-tooltip bottom>
-                                                <template v-slot:activator="{ on, attrs }">
-                                                    <span
-                                                        class="font-weight-bold"
-                                                        v-bind="attrs"
-                                                        v-on="on"
-                                                    >
-                                                        Subject: {{ subjects[group].name }}
-                                                    </span>
-                                                </template>
-                                                <span>
-                                                    ID: {{ subjects[group].id }} |
-                                                    Created: {{ shortDateString(subjects[group].created) }} |
-                                                    Modified: {{ shortDateString(subjects[group].modified) }}
-                                                </span>
-                                                </v-tooltip>
-                                            </td>
-                                        </template>
-                                        <!-- eslint-disable-next-line -->
-                                        <template v-slot:item.file="{ item }">
-                                            <span>{{ shortFileName(item.file) }}</span>
-                                        </template>
-                                    </v-data-table>
-                                </v-list-group>
-                            </v-list-item-group>
-                        </v-list>
-                    </div>
-                </v-list-item>
-                <br/>
-            </v-navigation-drawer>
-
-            <div :class="mini ?'pa-5 render-area maximize' :'pa-5 render-area'">
-                <span v-if="selectedDataObjects.length == 0">Select any number of data objects</span>
-                <template v-else>
-                <shape-viewer
-                    :data="renderData"
-                    :rows="rows"
-                    :columns="cols"
-                    :glyph-size="1.5"
-                />
-                </template>
-            </div>
-        </div>
-    </div>
-</template>
-
-<style>
-.context-card {
-    display: flex;
-    column-gap: 40px;
-    padding: 10px 20px;
-    width: 100%;
-    height: 80px;
-}
-.file-column {
-    width: 100px!important;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-.v-data-table td {
-    border-bottom: none !important;
-}
-.v-row-group__header {
-    background: none !important;
-    border-top: 1px solid white !important;
-}
-.content-area {
-    position: relative;
-    min-height: calc(100vh - 161px);
-}
-.render-area {
-    position: absolute;
-    display: flex;
-    justify-content: space-between;
-    left: 500px;
-    width: calc(100% - 500px);
-    height: 100%;
-}
-.render-area.maximize {
-    left: 60px;
-    width: calc(100% - 60px);
-}
-.render-area > * {
-    flex-grow: 1;
-}
-</style>
